+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 2020-06-29
 - add: name parser can be configured by a config. Keys: suffix, prefix, titles 
 - fix: campaign.group (groupGuid, group, groupId). Callback: campaignGroup
+- chg: the memo return value NOT text
 
 2020-06-28
 - fix: campaign action should call campaignContact